✨ feat: 添加 ImageRenderer 复用代码
