+++ README.md
@@ -123,7 +123,7 @@ Windows 参考(原项目推荐): https://www.jianshu.com/p/5015a477de3c
 | parser_disabled_platforms |  否   |            []            |                               全局禁止的解析，示例 parser_disabled_platforms=["bilibili", "douyin"] 表示禁止了哔哩哔哩和抖, 请根据自己需求填写["bilibili", "douyin", "kuaishou", "twitter", "youtube", "acfun", "tiktok", "weibo", "xiaohongshu"]                               |
 |    parser_render_type     |  否   |         "common"         |                                                                                             渲染器类型，可选 "default"(无图片渲染), "common"(PIL 通用图片渲染), "htmlkit"(htmlkit, 暂不可用)                                                                                              |
 |     parser_append_url     |  否   |          False           |                                                                                                                           是否在解析结果中附加原始URL                                                                                                                           |
-|    parser_custom_font     |  否   |           None           |                                                                           自定义渲染字体，配置字体文件名，并将字体文件放置于 localstore 生成的插件 data 目录下名,（如 ./data/nonebot_plugin_parser/）                                                                           |
+|    parser_custom_font     |  否   |           None           |                                                                           自定义渲染字体，配置字体文件名，并将字体文件放置于 localstore 生成的插件 data 目录下（如 ./data/nonebot_plugin_parser/）                                                                           |
 ## 🎉 使用
 ### 指令表
 |   指令   |         权限          | 需要@ | 范围  |   说明   |
